// packages/fe-enroll-permission-pc/src/pages/activity-management/detail/interface.d.ts
export interface ACTIVITY_DETAILS {
    code?: string
    sid?: number
    siteName?: string
    coverUrl?: string
    activityName?: string
    activityCatalogCode?: string
    activityCatalogName?: string
    activityForm?: number
    activityFormName?: string
    startTime?: string
    endTime?: string
    address?: string
    activityStatus?: number
    activityStatusName?: string
    activityIntroduce?: string
    relateProfessionStatus?: number
}